+++ tests/it/bitmap/mod.rs
@@ -1,2 +1,3 @@
 mod immutable;
 mod mutable;
+mod utils;
